+++ app/authenticate/page.tsx
@@ -1,0 +1,5 @@
+import Authenticate from '@/components/Authenticate';
+
+export default function AuthenticatePage() {
+  return <Authenticate />;
+}
