+++ app/layout.tsx
@@ -1,22 +1,24 @@
-import type { Metadata } from 'next';
-import { Inter } from 'next/font/google';
-import './globals.css';
+import Header from '@/components/Header';
+import StytchProvider from '@/components/StytchProvider';
+import type { ReactNode } from 'react';
+import './global.css';
 
-const inter = Inter({ subsets: ['latin'] });
-
-export const metadata: Metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
+type Props = {
+  children: ReactNode;
 };
 
-export default function RootLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
+export default function RootLayout({ children }: Props) {
   return (
-    <html lang='en'>
-      <body className={inter.className}>{children}</body>
-    </html>
+    <StytchProvider>
+      <html lang='en'>
+        <title>Financial Freedom</title>
+        <body>
+          <Header />
+          <main>
+            <div>{children}</div>
+          </main>
+        </body>
+      </html>
+    </StytchProvider>
   );
 }
