+++ app/profile/page.tsx
@@ -1,18 +1,20 @@
-import Login from '@/components/login';
+'use client';
+
+import Profile from '@/components/profile';
 import { useStytchUser } from '@stytch/nextjs';
-import { useRouter } from 'next/router';
+import { useRouter } from 'next/navigation';
 import { useEffect } from 'react';
 
-export default function LoginPage() {
+export default function ProfilePage() {
   const { isInitialized, user } = useStytchUser();
 
   const router = useRouter();
 
   useEffect(() => {
-    if (isInitialized && user) {
-      router.replace('/profile');
+    if (isInitialized && !user) {
+      router.replace('/');
     }
   }, [isInitialized, router, user]);
 
-  return <Login />;
+  return <Profile />;
 }
