+++ components/Authenticate.tsx
@@ -1,48 +1,40 @@
+'use client';
+
 import { useStytch, useStytchUser } from '@stytch/nextjs';
-import { useRouter } from 'next/router';
+import { useRouter, useSearchParams } from 'next/navigation';
 import { useEffect } from 'react';
 
 const OAUTH_TOKEN = 'oauth';
 const MAGIC_LINKS_TOKEN = 'magic_links';
 
-const Authenticate = () => {
+export default function Authenticate() {
   const { isInitialized, user } = useStytchUser();
   const stytch = useStytch();
   const router = useRouter();
+  const searchParams = useSearchParams();
 
   useEffect(() => {
     if (stytch && !user && isInitialized) {
-      const stytch_token_type = router?.query.stytch_token_type?.toString();
-      const token = router?.query.token?.toString();
+      const token = searchParams?.get('token');
+      const stytch_token_type = searchParams?.get('stytch_token_type');
 
       if (token && stytch_token_type === OAUTH_TOKEN) {
-        stytch.oauth.authenticate(token, {
-          session_duration_minutes: 60,
-        });
+        stytch.oauth.authenticate(token, { session_duration_minutes: 60 });
       } else if (token && stytch_token_type === MAGIC_LINKS_TOKEN) {
-        stytch.magicLinks.authenticate(token, {
-          session_duration_minutes: 60,
-        });
+        stytch.magicLinks.authenticate(token, { session_duration_minutes: 60 });
       }
     }
-  }, [
-    isInitialized,
-    router?.query.stytch_token_type,
-    router?.query.token,
-    stytch,
-    user,
-  ]);
+  }, [isInitialized, searchParams, stytch, user]);
 
   useEffect(() => {
     if (!isInitialized) {
       return;
     }
+
     if (user) {
       router.replace('/profile');
     }
   }, [isInitialized, router, user]);
 
   return null;
-};
-
-export default Authenticate;
+}
