+++ components/Header.tsx
@@ -1,0 +1,9 @@
+import Link from 'next/link';
+
+export default function Header() {
+  return (
+    <header>
+      <Link href='/'>Home</Link>
+    </header>
+  );
+}
