+++ components/profile.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { useStytch, useStytchSession, useStytchUser } from '@stytch/nextjs';
 
 export default function Profile() {
@@ -6,27 +8,25 @@ export default function Profile() {
   const { session } = useStytchSession();
 
   return (
-    <div className='card'>
+    <div>
       <h1>Profile</h1>
       <h2>User object</h2>
-      <pre className='code-block'>
-        <code>{JSON.stringify(user, null, 2)}</code>
+      <pre>
+        <code>{JSON.stringify(user, null, `\t`)}</code>
       </pre>
 
       <h2>Session object</h2>
-      <pre className='code-block'>
-        <code>{JSON.stringify(session, null, 2)}</code>
+      <pre>
+        <code>{JSON.stringify(session, null, `\t`)}</code>
       </pre>
       <p>
         You are logged in, and a Session has been created. The SDK stores the
         Session as a token and a JWT in the browser cookies as{' '}
-        <span className='code'>stytch_session</span> and{' '}
-        <span className='code'>stytch_session_jwt</span> respectively.
+        <span>stytch_session</span> and <span>stytch_session_jwt</span>{' '}
+        respectively.
       </p>
       {/* Revoking the session results in the session being revoked and cleared from browser storage. The user will return to Login.js */}
-      <button className='primary' onClick={() => stytch.session.revoke()}>
-        Log out
-      </button>
+      <button onClick={() => stytch.session.revoke()}>Log out</button>
     </div>
   );
 }
