+++ lib/urlUtils.ts
@@ -1,0 +1,7 @@
+export function getDomainFromWindow() {
+  if (typeof window !== 'undefined') {
+    return window.location.origin;
+  }
+
+  return null;
+}
